bathroom view updates
