+++ package.json
@@ -26,14 +26,15 @@
     "express-session": "^1.17.1",
     "galio-framework": "^0.6.3",
     "geolib": "^3.3.1",
+    "native-base": "^2.13.12",
     "nodemon": "^2.0.4",
     "passport": "^0.4.1",
     "passport-google-oauth": "^2.0.0",
-    "native-base": "^2.13.12",
     "pg": "^8.2.1",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
+    "react-native-animated-splash-screen": "^2.0.3",
     "react-native-elements": "^2.0.2",
     "react-native-gesture-handler": "~1.6.0",
     "react-native-google-places-autocomplete": "^1.7.1",
